@@ -1,6 +1,6 @@
 {
   "name": "dom-lib",
-  "version": "0.2.6",
+  "version": "1.0.0-beta.1",
   "description": "DOM helper library",
   "main": "lib/index.js",
   "directories": {
@@ -20,22 +20,31 @@
   },
   "files": [
     "CHANGELOG.md",
+    "src",
     "lib"
   ],
   "devDependencies": {
     "babel-core": "^6.7.6",
     "babel-eslint": "^7.2.1",
     "babel-loader": "^6.2.4",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-es2015-classes": "^6.14.0",
     "babel-plugin-transform-proto-to-assign": "^6.9.0",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-0": "^6.5.0",
+    "brfs": "^1.5.0",
     "chai": "^3.5.0",
     "es5-shim": "^4.1.14",
-    "eslint": "^2.8.0",
-    "eslint-plugin-babel": "^3.2.0",
-    "eslint-plugin-react": "^5.0.1",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-flowtype": "^2.40.1",
+    "eslint-plugin-import": "^2.6.1",
+    "eslint-plugin-json": "^1.2.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.3.0",
+    "flow-bin": "^0.68.0",
     "jquery": "^3.2.1",
     "karma": "^0.13.22",
     "karma-chrome-launcher": "^0.2.3",
