@@ -1,3 +1,5 @@
+// @flow
+
 import canUseDOM from './query/canUseDOM';
 
 /**
@@ -7,8 +9,7 @@ import canUseDOM from './query/canUseDOM';
  * `ReactWorker` dependency.
  */
 const ExecutionEnvironment = {
-
-  canUseDOM: canUseDOM,
+  canUseDOM,
   canUseWorkers: typeof Worker !== 'undefined',
   canUseEventListeners: canUseDOM && !!(window.addEventListener || window.attachEvent),
   canUseViewport: canUseDOM && !!window.screen,
