@@ -1,6 +1,8 @@
+// @flow
 
 import nativeRequestAnimationFrame from './nativeRequestAnimationFrame';
 import emptyFunction from '../utils/emptyFunction';
+
 let lastTime = 0;
 
 /**
@@ -9,14 +11,14 @@ let lastTime = 0;
  */
 const requestAnimationFrame =
   (nativeRequestAnimationFrame && nativeRequestAnimationFrame.bind(global)) ||
-  function (callback) {
+  (callback => {
     var currTime = Date.now();
     var timeDelay = Math.max(0, 16 - (currTime - lastTime));
     lastTime = currTime + timeDelay;
     return global.setTimeout(() => {
       callback(Date.now());
     }, timeDelay);
-  };
+  });
 
 // Works around a rare bug in Safari 6 where the first request is never invoked.
 requestAnimationFrame(emptyFunction);
