@@ -2,9 +2,9 @@ export * from './classNames';
 export * from './style';
 export * from './query';
 export * from './events';
+export * as animation from './animation';
 export transition from './transition';
 export translateDOMPositionXY from './transition/translateDOMPositionXY';
-export animation from './animation';
 export getVendorPrefixedName from './getVendorPrefixedName';
 export BrowserSupportCore from './BrowserSupportCore';
 export DOMMouseMoveTracker from './DOMMouseMoveTracker';
