@@ -1,6 +1,9 @@
+// @flow
 
-export default !!(
+const canUseDOM: boolean = !!(
   typeof window !== 'undefined' &&
   window.document &&
   window.document.createElement
 );
+
+export default canUseDOM;
