@@ -1,4 +1,6 @@
-export default (container, defaultContainer) => {
+// @flow
+
+export default (container: HTMLElement | Function, defaultContainer: HTMLElement): HTMLElement => {
   container = typeof container === 'function' ? container() : container;
   return container || defaultContainer;
 };
