@@ -1,7 +1,12 @@
 import getWindow from './getWindow';
 import getOffset from './getOffset';
 
-export default (node, client) => {
-  var win = getWindow(node);
-  return win ? win.innerHeight : client ? node.clientHeight : getOffset(node).height;
+export default (node: HTMLElement, client: HTMLElement): number => {
+  const win = getWindow(node);
+
+  if (win) {
+    return win.innerHeight;
+  }
+
+  return client ? node.clientHeight : getOffset(node).height;
 };
