@@ -1,10 +1,14 @@
+// @flow
+
 import canUseDOM from './canUseDOM';
 
-export default (recalc) => {
-  let size;
-  if (!size || recalc) {
+let size;
+
+export default (recalc: boolean): number | void => {
+  if (size === undefined || recalc) {
     if (canUseDOM) {
-      let scrollDiv = document.createElement('div');
+      const scrollDiv = document.createElement('div');
+      const body: any = document.body;
 
       scrollDiv.style.position = 'absolute';
       scrollDiv.style.top = '-9999px';
@@ -12,9 +16,9 @@ export default (recalc) => {
       scrollDiv.style.height = '50px';
       scrollDiv.style.overflow = 'scroll';
 
-      document.body.appendChild(scrollDiv);
+      body.appendChild(scrollDiv);
       size = scrollDiv.offsetWidth - scrollDiv.clientWidth;
-      document.body.removeChild(scrollDiv);
+      body.removeChild(scrollDiv);
     }
   }
 
