@@ -1,7 +1,20 @@
+// @flow
+
 import getWindow from './getWindow';
 import getOffset from './getOffset';
 
-export default (node, client) => {
-  var win = getWindow(node);
-  return win ? win.innerWidth : client ? node.clientWidth : getOffset(node).width;
+export default (node: HTMLElement, client: HTMLElement): number => {
+  const win = getWindow(node);
+
+  if (win) {
+    return win.innerWidth;
+  }
+
+  if (client) {
+    return node.clientWidth;
+  }
+
+  const offset = getOffset(node);
+
+  return offset ? offset.width : 0;
 };
