@@ -1,5 +1,9 @@
-export default (node) => {
-  return node === node.window ?
-    node : node.nodeType === 9 ?
-      node.defaultView || node.parentWindow : false;
+// @flow
+
+export default (node: any): window => {
+  if (node === node.window) {
+    return node;
+  }
+
+  return node.nodeType === 9 ? node.defaultView || node.parentWindow : null;
 };
