@@ -1,6 +1,7 @@
+// @flow
+
 import { camelize } from '../utils/stringFormatter';
+
 const msPattern = /^ms-/;
 
-export default (string) => {
-  return camelize(string.replace(msPattern, 'ms-'));
-};
+export default (name: string) => camelize(name.replace(msPattern, 'ms-'));
