@@ -1,7 +1,5 @@
-
 import { hyphenate } from '../utils/stringFormatter';
+
 const msPattern = /^ms-/;
 
-export default (string) => {
-  return hyphenate(string).replace(msPattern, '-ms-');
-};
+export default string => hyphenate(string).replace(msPattern, '-ms-');
