@@ -1,21 +1,14 @@
 function makeEmptyFunction(arg) {
-  return function () {
-    return arg;
-  };
+  return () => arg;
 }
 
-function emptyFunction() { }
+function emptyFunction() {}
 
 emptyFunction.thatReturns = makeEmptyFunction;
 emptyFunction.thatReturnsFalse = makeEmptyFunction(false);
 emptyFunction.thatReturnsTrue = makeEmptyFunction(true);
 emptyFunction.thatReturnsNull = makeEmptyFunction(null);
-emptyFunction.thatReturnsThis = function () {
-  return this;
-};
-emptyFunction.thatReturnsArgument = function (arg) {
-  return arg;
-};
-
+emptyFunction.thatReturnsThis = () => this;
+emptyFunction.thatReturnsArgument = arg => arg;
 
 export default emptyFunction;
