@@ -1,13 +1,14 @@
+/* eslint-disable */
 
 /**
  * @example
  * underscoreName('getList');
  * => get_list
  */
+
 export function underscore(string) {
   return string.replace(/([A-Z])/g, '_$1').toLowerCase();
 }
-
 
 /**
  * @example
@@ -15,7 +16,7 @@ export function underscore(string) {
  * => fontSize
  */
 export function camelize(string) {
-  return string.replace(/\-(\w)/g, function (char) {
+  return string.replace(/\-(\w)/g, char => {
     return char.slice(1).toUpperCase();
   });
 }
@@ -29,7 +30,6 @@ export function hyphenate(string) {
   return string.replace(/([A-Z])/g, '-$1').toLowerCase();
 }
 
-
 /**
  * @example
  * merge('{0} - A front-end {1} ','Suite','framework');
@@ -38,9 +38,9 @@ export function hyphenate(string) {
 export function merge(pattern) {
   var pointer = 0,
     i;
-  for (i = 1; i < arguments.length; i++) {
-    pattern = pattern.split('{' + pointer + '}').join(arguments[i]);
-    pointer++;
+  for (i = 1; i < arguments.length; i += 1) {
+    pattern = pattern.split(`{${pointer}}`).join(arguments[i]);
+    pointer += 1;
   }
   return pattern;
 }
