@@ -7,17 +7,21 @@
  * @param options
  * @returns {throttled}
  */
+
+/* eslint-disable */
 export default function throttle(func, wait, options) {
   var context, args, result;
   var timeout = null;
   var previous = 0;
-  var _now = Date.now || function () {
-    return new Date().getTime();
-  };
+  var _now =
+    Date.now ||
+    function() {
+      return new Date().getTime();
+    };
   if (!options) {
     options = {};
   }
-  var later = function () {
+  var later = function() {
     previous = options.leading === false ? 0 : _now();
     timeout = null;
     result = func.apply(context, args);
@@ -25,7 +29,7 @@ export default function throttle(func, wait, options) {
       context = args = null;
     }
   };
-  return function () {
+  return function() {
     var now = _now();
     if (!previous && options.leading === false) {
       previous = now;
